[Refactor] Users entity들을 entities 폴더로 묶으로써 변경된 사항들 추가
