+++ back/web-app/src/users/dto/create-member-user.dto.ts
@@ -1,4 +1,4 @@
-import { User } from "../user.entity";
+import { User } from "../entities/user.entity";
 
 export class CreateMemberUserDto { // 영문, 숫자, 길이 체크 pipe
   public id: string;
