+++ back/web-app/src/users/dto/create-oauth-user.dto.ts
@@ -1,4 +1,4 @@
-import { User } from "../user.entity";
+import { User } from "../entities/user.entity";
 
 export class CreateOauthUserDto {
   public user: User;
