+++ back/web-app/src/users/user.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { UserService } from './user.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from './user.entity';
-import { OauthUser } from './oauth-user.entity';
-import { MemberUser } from './member-user.entity';
+import { User } from './entities/user.entity';
+import { OauthUser } from './entities/oauth-user.entity';
+import { MemberUser } from './entities/member-user.entity';
 
 @Module({
   imports: [
